requirements.txt
streamlit
pandas
matplotlib
psycopg2-binary
polars
numpy
